Refactor: Update module files locaton
