+++ src/pages/404.html
@@ -54,7 +54,7 @@
     <link rel="stylesheet" href="/styles/404.css">
 
     <!-- Scripts -->
-    <script src="/scripts/common.js"></script>
+    <script src="/scripts/common.js" type="module"></script>
     <noscript>
         <meta http-equiv="refresh" content="0; URL=/noscript/">
     </noscript>
