+++ src/pages/eqhistory/index.html
@@ -67,7 +67,7 @@
     <!-- Scripts -->
     <script src="https://unpkg.com/axios/dist/axios.min.js"></script>
 
-    <script src="/scripts/common.js"></script>
+    <script src="/scripts/common.js" type="module"></script>
     <script src="/scripts/main.mjs" type="module"></script>
 
     <!-- Google tag (gtag.js) -->
