+++ src/pages/help/index.html
@@ -53,7 +53,7 @@
     <link rel="stylesheet" href="./styles/index.css">
 
     <!-- Scripts -->
-    <script src="/scripts/common.js"></script>
+    <script src="/scripts/common.js" type="module"></script>
     <noscript>
         <meta http-equiv="refresh" content="0; URL=/noscript/">
     </noscript>
