+++ src/pages/help/sound-setting/index.html
@@ -55,7 +55,7 @@
 
 
     <!-- Scripts -->
-    <script src="/scripts/common.js"></script>
+    <script src="/scripts/common.js" type="module"></script>
     <noscript>
         <meta http-equiv="refresh" content="0; URL=/noscript/">
     </noscript>
