+++ src/pages/help/support-browser/index.html
@@ -56,7 +56,7 @@
     <link rel="stylesheet" href="./styles/index.css">
 
     <!-- Scripts -->
-    <script src="/scripts/common.js"></script>
+    <script src="/scripts/common.js" type="module"></script>
     <noscript>
         <meta http-equiv="refresh" content="0; URL=/noscript/">
     </noscript>
