+++ src/pages/index.html
@@ -68,8 +68,8 @@
     <script src="https://cdn.jsdelivr.net/npm/maplibre-gl@^5.3.0/dist/maplibre-gl.js"></script>
     <script src="https://cdn.jsdelivr.net/npm/@turf/turf@7/turf.min.js"></script>
 
-    <script src="/scripts/common.mjs" type="module"></script>
-    <script src="/scripts/main.mjs" type="module"></script>
+    <script src="/scripts/common.js" type="module"></script>
+    <script src="/scripts/main.js" type="module"></script>
 
     <!-- Google tag (gtag.js) -->
     <script async src="https://www.googletagmanager.com/gtag/js?id=G-TWBJ0TEFRJ"></script>
