+++ src/pages/scripts/common.js
