+++ src/pages/scripts/ydits-web/services/debug-logs/debug-logs.js
@@ -8,7 +8,7 @@
  *
  */
 
-import { Service } from "../../../service.mjs";
+import { Service } from "../../../packages/app-creater/service.js";
 
 /**
  * デバッグログを管理する。
