+++ src/pages/scripts/ydits-web/services/eew/eew.js
@@ -8,7 +8,7 @@
  *
  */
 
-import { Service } from "../../../service.mjs";
+import { Service } from "../../../packages/app-creater/service.js";
 
 /**
  * 緊急地震速報を扱う。
