+++ src/pages/scripts/ydits-web/services/intensity/intensity.js
