+++ src/pages/scripts/ydits-web/services/local-storage/local-storage.js
@@ -8,7 +8,7 @@
  *
  */
 
-import { Service } from "../../../service.mjs";
+import { Service } from "../../../packages/app-creater/service.js";
 
 export class LocalStorage extends Service {
     constructor(app) {
