+++ src/pages/scripts/ydits-web/services/push-notify/push-notify.js
@@ -8,7 +8,7 @@
  *
  */
 
-import { Service } from "../../../service.mjs"
+import { Service } from "../../../packages/app-creater/service.js";
 
 /**
  * プッシュ通知を扱う。
