+++ src/pages/scripts/ydits-web/services/service-worker/service-worker.js
@@ -8,7 +8,7 @@
  *
  */
 
-import { Service } from "../../service.mjs";
+import { Service } from "../../../packages/app-creater/service.js";
 
 /**
  * サービスワーカーを管理する
